chore: rename client ping API to heartbeatwe
